// apps/web/src/generated/folders.ts
// Generated by crudder. DO NOT EDIT.

export interface CreateFolderRequest {
    name: string;
    color?: string;
    parent_id?: string;
}

export interface Folder {
    id: number;
    public_id: string;
    user_id: number;
    name: string;
    color?: string;
    parent_id?: number;
    collapsed: boolean;
    created_at: string;
    updated_at: string;
}

export interface FolderResponse {
    folder: Folder;
}

export interface FoldersResponse {
    folders: Folder[];
}

export interface UpdateFolderRequest {
    name?: string;
    color?: string;
    collapsed?: boolean;
    parent_id?: string;
}